+++ extensions/yii/bootstrap/CHANGELOG.md
@@ -7,8 +7,12 @@ Yii Framework 2 bootstrap extension Change Log
 - Enh #1474: Added option to make NavBar 100% width (cebe)
 - Enh #1553: Only add navbar-default class to NavBar when no other class is specified (cebe)
 - Enh #1601: Added support for tagName and encodeLabel parameters in ButtonDropdown (omnilight)
+- Enh #1881: innerContainerOptions added to NavBar (creocoder)
+- Enh #1881: special tag attribute added to NavBar options (creocoder)
 - Chg #1459: Update Collapse to use bootstrap 3 classes (tonydspaniard)
 - Chg #1820: Update Progress to use bootstrap 3 markup (samdark)
+- Chg #1881: NavBar padded renamed to renderInnerContainer (creocoder)
+- Chg #1881: NavBar collapseOptions added (creocoder)
 
 2.0.0 alpha, December 1, 2013
 -----------------------------